chore: update Storybook dependency from @storybook/react-vite to @storybook/experimental-nextjs-vite version 8.6.12
