+++ libs/shared-ui/src/components/index.ts
@@ -1,3 +1,7 @@
+export * from "./filter-drawer";
+export * from "./movie-card";
 export * from "./seen-chips";
 export * from "./shared";
-export * from "./filter-drawer";
+export * from "./user-flag-button";
+export * from "./user-flag-chip";
+
