+++ libs/shared-ui/src/components/user-flag-chip.stories.tsx
@@ -1,0 +1,21 @@
+import type { Meta, StoryObj } from "@storybook/react";
+import { UserFlagChip } from "./user-flag-chip";
+
+const meta: Meta<typeof UserFlagChip> = {
+  component: UserFlagChip,
+  title: "UserFlagChip",
+};
+export default meta;
+type Story = StoryObj<typeof UserFlagChip>;
+
+export const Watchagain: Story = {
+  args: {
+    type: "Watchagain"
+  },
+};
+
+export const Favorite: Story = {
+  args: {
+    type: "Favorite"
+  }
+};
