+++ libs/shared-ui/src/components/user-flag-chip.tsx
@@ -9,14 +9,12 @@ export interface UserFlagChipProperties {
 }
 export const UserFlagChip = ({ type, active, loading }: UserFlagChipProperties) => {
   return (
-    <>
-      <Chip className={`text-left w-full ${loading ? "animate-pulse" : ""}`} color={active ? "warning" : "default"}>
-        {type === "Favorite" ? (
-          active ? <HeartFilled /> : <HeartOutlined />
-        ) : (
-          active ? <FlagFilled /> : <FlagOutlined />
-        )}
-      </Chip>
-    </>
+    <Chip className={`text-left w-full ${loading ? "animate-pulse" : ""}`} color={active ? "warning" : "default"}>
+      {type === "Favorite" ? (
+        active ? <HeartFilled /> : <HeartOutlined />
+      ) : (
+        active ? <FlagFilled /> : <FlagOutlined />
+      )}
+    </Chip>
   );
 };
