+++ libs/shared-ui/src/interfaces/index.ts
@@ -1,0 +1,4 @@
+export * from "./movie-db-session";
+export * from "./movie";
+export * from "./seen-date-dto";
+export * from "./user-flags-dto";
