+++ libs/shared-ui/src/interfaces/movie-db-session.ts
@@ -1,0 +1,6 @@
+export interface MoviesDbSession {
+  userName: string;
+  ownerId: number;
+  eMail: string;
+  avatarUrl: string;
+}
