+++ libs/shared-ui/src/interfaces/movie.ts
@@ -1,0 +1,12 @@
+// Define the interface for a single movie
+export interface Movie {
+  id: string;
+  title: string;
+  subtitle: string;
+  diskid?: string; // Optional
+  mediaType: string;
+  genres: string[];
+  ownerid: string;
+  istv: boolean;
+  plot: string;
+}
