+++ libs/shared-ui/src/interfaces/seen-date-dto.ts
@@ -1,0 +1,4 @@
+export interface SeenDateDTO {
+  movieId: string;
+  dates: string[];
+}
