+++ libs/shared-ui/src/interfaces/user-flags-dto.ts
@@ -1,0 +1,5 @@
+export interface UserFlagsDTO {
+  movieId: string;
+  isWatchAgain: boolean;
+  isFavorite: boolean;
+}
