+++ package.json
@@ -55,7 +55,7 @@
     "@storybook/addon-interactions": "^8.4.6",
     "@storybook/core-server": "^8.4.6",
     "@storybook/jest": "^0.2.3",
-    "@storybook/react-vite": "^8.4.6",
+    "@storybook/experimental-nextjs-vite": "^8.6.12",
     "@storybook/test-runner": "^0.22.0",
     "@storybook/testing-library": "^0.2.2",
     "@swc-node/register": "~1.10.0",
